requirements.txt
absl-py==2.3.1
altgraph==0.17.5
astunparse==1.6.3
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.3
flask-cors==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.76.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pefile==2024.8.26
pillow==12.1.0
protobuf==6.33.4
pycryptodome==3.23.0
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
requests==2.32.5
rich==14.2.0
sib-api-v3-sdk==7.6.0
six==1.17.0
SQLAlchemy==2.0.49
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
